requirements.txt
Flask
gunicorn
typing
sklearn
pandas
numpy 
PyYAML
-e .
